Add test to reproduce race cond. in watchDir/close
If a call to Gaze#close is made right between when Gaze#watchDir's
listener fires and the timeout executes, the watcher continues to watch
for changes.

This requires a separate node process to keep making changes to the file
as a single node process is single threaded and IO-bound so it would
never encounter this scenario.